+++ extensions/readwise-to-tana/src/book.tsx
@@ -4,6 +4,7 @@ import Highlight from './highlight'
 import React from 'react'
 import Handlebars from 'handlebars'
 import { cleanTitle, formatDate } from './utils'
+import { ifeq } from './helpers'
 
 type BookProps = {
   id: string
@@ -48,17 +49,24 @@ export default function Book({ id, template }: BookProps) {
   }
 
   const h = Handlebars.compile(template)
-  const allHighlights = h({
-    ...book,
-    title: cleanTitle(book.title),
-    highlights: highlights.map((highlight) => ({
-      ...highlight,
-      updated: formatDate(highlight.updated),
-      highlighted_at: formatDate(highlight.highlighted_at),
-      note: (highlight.note ?? '').split('\n').filter((note) => note),
-      tags: highlight.tags.map(({ name }) => name).join(', '),
-    })),
-  })
+  const allHighlights = h(
+    {
+      ...book,
+      title: cleanTitle(book.title),
+      highlights: highlights.map((highlight) => ({
+        ...highlight,
+        updated: formatDate(highlight.updated),
+        highlighted_at: formatDate(highlight.highlighted_at),
+        note: (highlight.note ?? '').split('\n').filter((note) => note),
+        tags: highlight.tags.map(({ name }) => name).join(', '),
+      })),
+    },
+    {
+      helpers: {
+        ifeq,
+      },
+    }
+  )
 
   const handleCopyAll = async () => {
     const currentTime = new Date().toISOString()